requirements.txt
# Core HTTP and API
httpx==0.27.0
fastapi==0.110.0
uvicorn==0.29.0
websockets==12.0
gunicorn==21.2.0

# Configuration and Validation
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Security and Crypto
cryptography==42.0.5
python-jose[cryptography]==3.3.0

# Redis (async)
redis[asyncio]>=4.5.0

# Advanced features
watchdog==4.0.0
weasyprint==61.2

# Utilities
rich==13.7.1
aiohttp==3.9.3

# Development, Testing and Linting
pytest==8.0.2
pytest-asyncio==0.23.5
fakeredis>=2.21.0
locust>=2.15.0
prometheus-client==0.20.0
coverage==7.4.3
ruff==0.3.2
black==24.2.0
